Simplify API by splitting `/api/comments` into two routes
